Move helper function to base
