+++ openregistry/lots/loki/tests/blanks/lot_blanks.py
@@ -19,7 +19,13 @@ from openregistry.lots.loki.constants import (
     DEFAULT_DUTCH_STEPS,
     RECTIFICATION_PERIOD_DURATION
 )
-
+from openregistry.lots.loki.tests.base import (
+    create_single_lot,
+    check_patch_status_200,
+    check_patch_status_403,
+    add_decisions,
+    add_auctions
+)
 
 ROLES = ['lot_owner', 'Administrator', 'concierge', 'convoy', 'chronograph']
 STATUS_BLACKLIST = create_blacklist(STATUS_CHANGES, LOT_STATUSES, ROLES)
@@ -92,74 +98,6 @@ def add_decisions(self, lot):
     self.assertEqual(response.content_type, 'application/json')
     self.assertEqual(response.json['data']['decisions'], data_with_decisions['decisions'])
 
-
-def add_auctions(self, lot, access_header):
-    response = self.app.get('/{}/auctions'.format(lot['id']))
-    auctions = sorted(response.json['data'], key=lambda a: a['tenderAttempts'])
-    english = auctions[0]
-    second_english = auctions[1]
-
-    response = self.app.patch_json(
-        '/{}/auctions/{}'.format(lot['id'], english['id']),
-        params={'data': auction_english_data}, headers=access_header)
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(response.content_type, 'application/json')
-
-    response = self.app.patch_json(
-        '/{}/auctions/{}'.format(lot['id'], second_english['id']),
-        params={'data': auction_second_english_data}, headers=access_header)
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(response.content_type, 'application/json')
-
-
-
-def check_patch_status_200(self, path, lot_status, headers=None):
-    response = self.app.patch_json(path,
-                                   headers=headers,
-                                   params={'data': {'status': lot_status}})
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(response.content_type, 'application/json')
-    self.assertEqual(response.json['data']['status'], lot_status)
-
-
-def check_patch_status_403(self, path, lot_status, headers=None):
-
-    # Check if response.status is forbidden, when you try to change status to incorrect
-    # 'data' should be {'data': {'status': allowed_status}}
-    response = self.app.patch_json(path,
-                                   params={'data': {'status': lot_status}},
-                                   headers=headers,
-                                   status=403)
-    self.assertEqual(response.status, '403 Forbidden')
-    self.assertEqual(response.content_type, 'application/json')
-    self.assertEqual(response.json['status'], 'error')
-
-
-def create_single_lot(self, data, status=None):
-    response = self.app.post_json('/', {"data": data})
-    self.assertEqual(response.status, '201 Created')
-    self.assertEqual(response.content_type, 'application/json')
-    self.assertEqual(response.json['data']['status'], 'draft')
-    self.assertEqual(len(response.json['data']['auctions']), 3)
-    token = response.json['access']['token']
-    lot_id = response.json['data']['id']
-
-    if status:
-        fromdb = self.db.get(lot_id)
-        fromdb = self.lot_model(fromdb)
-
-        fromdb.status = status
-        fromdb.store(self.db)
-
-        response = self.app.get('/{}'.format(lot_id))
-        self.assertEqual(response.status, '200 OK')
-        self.assertEqual(response.json['data']['id'], lot_id)
-        self.assertEqual(response.json['data']['status'], status)
-        new_json = deepcopy(response.json)
-        new_json['access'] = {'token': token}
-        return new_json
-
-    return response
 
 
 def auction_autocreation(self):
